+++ CHANGELOG.md
@@ -9,6 +9,12 @@ Translations: [简体中文](CHANGELOG.zh.md)
      prompt for upgrade.
 > 3. Reference [《Migration Documentation》](docs/migrate.md) migrating from 3.x to 4.x
 
+# new
+
+* fix: Fixed a bug that caused the App to crash or image loading failure when
+  ThumbnailMemoryCacheStateImage to encounter memory cache without
+  ImageInfo. [#259](https://github.com/panpf/sketch/issues/259)
+
 # 4.2.0
 
 > [!NOTE]